adds check for duplicate color assignment
