+++ maps/tests/bad_invalid_rgb4.cub
@@ -6,7 +6,7 @@ EA ./textures/test3.png
 C 0,50,100
 F 50,255,50
 
-C 10,150,100
+
