+++ src/handle_args.c
@@ -6,7 +6,7 @@
 /*   By: ktoivola <ktoivola@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/14 14:34:01 by ktoivola          #+#    #+#             */
-/*   Updated: 2025/01/14 14:34:04 by ktoivola         ###   ########.fr       */
+/*   Updated: 2025/01/15 13:48:11 by ktoivola         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,6 +21,26 @@
 	}
 	return (0);
 } */
+
+int	validate_arguments(int argc, char **argv)
+{
+	size_t	len;
+	int		fd;
+
+	if (argc != 2)
+		handle_error(ERROR_CMD_COUNT_ERROR);
+	len = ft_strlen(argv[1]);
+	if (len < 5 || ft_strncmp(argv[1] + len - 4, ".cub", 4) != 0)
+		handle_error(ERROR_INVALID_FILE_NAME);
+	if (ft_strncmp(argv[1], "maps/valid/", 11) != 0
+		&& ft_strncmp(argv[1], "maps/unvalid/", 13) != 0)
+		handle_error(ERROR_INVALID_FILE_NAME);
+	fd = open(argv[1], O_RDONLY);
+	if (fd == -1)
+		handle_error(ERROR_OPEN_ERROR);
+	close(fd);
+	return (0);
+}
 
 static int	check_file_format(const char *filename)
 {
